Revise Scenario structure
