+++ src/deflex/__init__.py
@@ -1,4 +1,4 @@
 __version__ = "0.2.1dev"
 
-from .scenario_tools import DeflexScenario  # noqa: F401
-from .scenario_tools import Scenario  # noqa: F401
+from .scenario import DeflexScenario  # noqa: F401
+from .scenario import Scenario  # noqa: F401
