+++ tests/test_scenario_creator.py
@@ -23,7 +23,7 @@ from scenario_builder import powerplants
 from deflex import __file__ as dfile
 from deflex import config as cfg
 from deflex import scenario_creator
-from deflex import scenario_tools as st
+from deflex import nodes as st
 from deflex.geometries import deflex_power_lines
 from deflex.geometries import deflex_regions
